cleanup codebase
