+++ app/auth/verification/[token]/page.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { VerificationForm } from "@/components/auth/verification/verification-form";
+
+const VerifyEmailPage = async ({ params }: { params: { token: string } }) => {
+  const token = (await params).token;
+
+  return <VerificationForm token={token} />;
+};
+
+export default VerifyEmailPage;
