+++ prisma/schema.prisma
@@ -100,7 +100,6 @@ model PasswordResetToken {
   id      String   @id @default(cuid())
   email   String
   token   String   @unique
-  code    String   @map("reset_code")
   expires DateTime
 
   @@unique([email, token])
